Move dialogs and custom controls under gui folder.
