+++ src/gui/CMakeLists.txt
@@ -1,1 +1,4 @@
+add_subdirectory(controls)
+add_subdirectory(dialogs)
 add_subdirectory(util)
+
