+++ src/gui/controls/plot.cpp
