+++ src/gui/controls/plot_scalar.cpp
@@ -19,8 +19,8 @@
 //
 //==========================================================================
 #include <string.h>
-#include "wx/wx.h"
-#include "main.h"
+
+#include <wx/wx.h>
 #include "plot_scalar.h"
 
 BEGIN_EVENT_TABLE(PlotScalar, PlotPanel)
