+++ src/gui/controls/plot_scalar.h
