+++ src/gui/controls/plot_scatter.cpp
@@ -19,9 +19,11 @@
 //
 //==========================================================================
 #include <string.h>
-#include "wx/wx.h"
-#include "plot_scatter.h"
+
+#include <wx/wx.h>
 #include <wx/graphics.h>
+
+#include "plot_scatter.h"
 
 BEGIN_EVENT_TABLE(PlotScatter, PlotPanel)
     EVT_PAINT           (PlotScatter::OnPaint)
