+++ src/gui/controls/plot_scatter.h
@@ -23,7 +23,7 @@
 
 #include "comp.h"
 #include "plot.h"
-#include "defines.h"
+#include "../../defines.h"
 
 #define PLOT_SCATTER_MODE_SCATTER            0
 #define PLOT_SCATTER_MODE_EYE                1
