+++ src/gui/dialogs/dlg_filter.cpp
