+++ src/gui/dialogs/dlg_filter.h
