+++ src/gui/dialogs/dlg_options.h
