+++ src/gui/dialogs/dlg_ptt.cpp
