+++ src/main.h
@@ -72,18 +72,12 @@
 #include "modem_stats.h"
 
 #include "topFrame.h"
-#include "dlg_easy_setup.h"
-#include "dlg_ptt.h"
-#include "dlg_options.h"
-#include "plot.h"
-#include "plot_scalar.h"
-#include "plot_scatter.h"
-#include "plot_waterfall.h"
-#include "plot_spectrum.h"
+#include "gui/controls/plot.h"
+#include "gui/controls/plot_scalar.h"
+#include "gui/controls/plot_scatter.h"
+#include "gui/controls/plot_waterfall.h"
+#include "gui/controls/plot_spectrum.h"
 #include "sndfile.h"
-#include "dlg_audiooptions.h"
-#include "dlg_filter.h"
-#include "dlg_options.h"
 #include "freedv_reporter.h"
 #include "sox_biquad.h"
 #include "comp_prim.h"
